+++ app/composables/usePlatformApi.ts
@@ -4,7 +4,7 @@ import { getPlatformApi } from '~/storage/platformRouter'
 export function usePlatformApi() {
   const { data: appConfigDisk } = useAppConfigDisk()
   const platformApi = computed(() =>
-    getPlatformApi(appConfigDisk.value.applicationType),
+    getPlatformApi(appConfigDisk.value.storageType),
   )
 
   return {