+++ app/config/loader.ts
@@ -2,7 +2,7 @@ import yaml from 'yaml'
 import rawDefaultConfig from './default.yaml?raw'
 import { parseAppConfig, type AppConfig } from './parseAppConfig'
 
-export type { AppConfig, ApplicationType } from './parseAppConfig'
+export type { AppConfig, StorageType } from './parseAppConfig'
 export { parseAppConfig } from './parseAppConfig'
 
 const DEFAULT_CONFIG = parseAppConfig(yaml.parse(rawDefaultConfig) as unknown)